Moved value converter include out of internal folder.
